app/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Flask-Limiter==3.5.0
gunicorn==21.2.0
google-cloud-firestore==2.13.1
google-cloud-logging==3.8.0
Werkzeug==2.3.7
